chore: split containers-api tests
