+++ packages/amplify-e2e-tests/src/__tests__/containers-api-2.test.ts
@@ -1,0 +1,45 @@
+import {
+  addRestContainerApi,
+  amplifyConfigureProject,
+  amplifyPushWithoutCodegen,
+  createNewProjectDir,
+  deleteProject,
+  deleteProjectDir,
+  initJSProjectWithProfile,
+  getProjectMeta,
+  modifyRestAPI,
+} from 'amplify-category-api-e2e-core';
+import fetch from 'node-fetch';
+import { getAWSExports } from '../aws-exports/awsExports';
+
+async function setupAmplifyProject(cwd: string) {
+  await amplifyConfigureProject({
+    cwd,
+    enableContainers: true,
+  });
+}
+
+describe('amplify api add', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('containers');
+  });
+
+  afterEach(async () => {
+    await deleteProject(projRoot);
+    deleteProjectDir(projRoot);
+  });
+
+  it('init project, enable containers and add multicontainer api push, edit and push', async () => {
+    const envName = 'devtest';
+    const apiName = 'containermodifyapi';
+    await initJSProjectWithProfile(projRoot, { name: 'multicontainer', envName });
+    await setupAmplifyProject(projRoot);
+    await addRestContainerApi(projRoot, { apiName });
+    await amplifyPushWithoutCodegen(projRoot);
+    const meta = await getProjectMeta(projRoot);
+    const api = Object.keys(meta.api)[0];
+    modifyRestAPI(projRoot, api);
+    await amplifyPushWithoutCodegen(projRoot);
+  });
+});
